+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v0.56.0
 
   - Add `Falcon::Environment::Cluster` and `Falcon::Service::Cluster` for running workers with independently bound endpoints.
   - Add `Falcon::Listener` to describe bound listeners shared by regular server workers or owned by cluster workers.